Ejercicio 5 del proyecto

+++ app/src/main/java/com/cifo/rgonzalezgall/mybooks/BookDetailActivity.java
@@ -55,8 +55,8 @@ public class BookDetailActivity extends AppCompatActivity {
             // using a fragment transaction.
             Bundle arguments = new Bundle();
             //Call to the fragment activity passing data with a Bundle
-            arguments.putInt(BookDetailFragment.ARG_ITEM_ID,
-                    getIntent().getIntExtra(BookDetailFragment.ARG_ITEM_ID, -1));
+            arguments.putLong(BookDetailFragment.ARG_ITEM_ID,
+                    getIntent().getLongExtra(BookDetailFragment.ARG_ITEM_ID, -1));
             BookDetailFragment fragment = new BookDetailFragment();
             fragment.setArguments(arguments);
             getSupportFragmentManager().beginTransaction()
